@@ -14,7 +14,9 @@ description: try to wirte all widget
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.0.0"
+
 
 dependencies:
   flutter:
@@ -22,25 +24,29 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.6
+  fl_chart: ^0.66.0
   english_words: ^4.0.0
   photo_view: ^0.14.0
-  cached_network_image: ^3.2.1
-  charts_flutter: ^0.12.0
-  fl_chart: ^0.55.0
-  provider: ^6.0.3
-  flutter_riverpod: ^2.0.0-dev.9
-  fconsole: ^2.3.0
-  logger: ^1.1.0
-  image: ^4.0.13
+  cached_network_image: ^3.3.1
+  flutter_riverpod: ^2.4.10
+  riverpod_annotation: ^2.3.4
+  fconsole: ^2.6.1
+  logger: ^2.0.2+1
+  image: ^4.1.6
   wrapper: ^1.0.2
-  sliver_tools: ^0.2.7
+  provider: ^6.1.1
+  sliver_tools: ^0.2.12
   flutter_sticky_header: ^0.6.5
 #  image: ^3.1.3
-  flutter_widget_from_html_core: ^0.8.5
+  flutter_widget_from_html_core: ^0.14.11
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  build_runner:
+  custom_lint:
+  riverpod_generator: ^2.3.9
+  riverpod_lint: ^2.3.7
 
 
 # For information on the generic Dart part of this file, see the
