@@ -29,108 +29,68 @@ class _ButtonWidgetState extends State<ButtonWidget> {
       ),
       body: ListView(
         children: <Widget>[
-          FlatButton(
+          TextButton(
             onPressed: null,
             child: Text("被禁用的button"),
-            disabledColor: Colors.blueGrey,
           ),
-          FlatButton(
+          TextButton(
             onPressed: () {
               print("点击了Button");
             },
             child: Text("可点击的button"),
-            color: Colors.tealAccent,
+
           ),
-          FlatButton.icon(
+          TextButton.icon(
             onPressed: () {
               print("点了了带有icon的FlatButton");
             },
             icon: Icon(Icons.print),
             label: Text("带有icon的FlatButton"),
           ),
-          FlatButton(
+          TextButton(
               // 文本内容
               child: Text("自定义按钮", semanticsLabel: 'FLAT BUTTON 2'),
-              // 按钮颜色
-              color: Colors.lightGreenAccent,
-              // 按钮亮度
-              colorBrightness: Brightness.dark,
-              // 高亮时的背景色
-              //highlightColor: Colors.yellow,
-              // 失效时的背景色
-              disabledColor: Colors.grey,
-              // 该按钮上的文字颜色，但是前提是不设置字体自身的颜色时才会起作用
-              textColor: Colors.white,
-              // 按钮失效时的文字颜色，同样的不能使用文本自己的样式或者颜色时才会起作用
-              disabledTextColor: Colors.grey,
-              // 按钮主题,主要用于与ButtonTheme和ButtonThemeData一起使用来定义按钮的基色,RaisedButton，FlatButton，OutlineButton，它们是基于环境ButtonTheme配置的
-              //ButtonTextTheme.accent，使用模版颜色的;ButtonTextTheme.normal，按钮文本是黑色或白色取决于。ThemeData.brightness;ButtonTextTheme.primary，按钮文本基于。ThemeData.primaryColor.
-              textTheme: ButtonTextTheme.normal,
-              // 按钮内部,墨汁飞溅的颜色,点击按钮时的渐变背景色，当你不设置高亮背景时才会看的更清楚
-              splashColor: Colors.deepPurple,
-              // 抗锯齿能力,抗锯齿等级依次递增,none（默认),hardEdge,antiAliasWithSaveLayer,antiAlias
-              clipBehavior: Clip.antiAlias,
-              padding: EdgeInsets.only(
-                  bottom: 5.0, top: 5.0, left: 30.0, right: 30.0),
-              shape: Border.all(
-                // 设置边框样式
-                color: Colors.grey,
-                width: 2.0,
-                style: BorderStyle.solid,
+              style: ButtonStyle(
+                //定义样式，这里设置颜色不起作用
+                textStyle: MaterialStateProperty.all(
+                  TextStyle(fontSize: 18,color: Colors.red)
+                ),
+                padding:MaterialStateProperty.all(
+                  EdgeInsets.only(
+                      bottom: 5.0, top: 5.0, left: 30.0, right: 30.0),
+                ),
+                foregroundColor: MaterialStateProperty.resolveWith((states)  {
+                  if(states.contains(MaterialState.pressed)){
+                     return Colors.blue;
+                  }
+                  return Colors.grey;
+                }),
+                backgroundColor: MaterialStateProperty.resolveWith((states)  {
+                  if(states.contains(MaterialState.pressed)){
+                    return Colors.blue[200];
+                  }
+                  return null;
+                }),
+                overlayColor: MaterialStateProperty.all(Colors.yellow),
+                side: MaterialStateProperty.all(BorderSide(color:Colors.grey,width: 1) ),
+                shape:MaterialStateProperty.all(StadiumBorder())
               ),
-              // FlatButton 的点击事件
-              onPressed: () {
-                // Perform some action
-              },
-              // 改变高亮颜色回掉函数，一个按钮会触发两次，按下后改变时触发一次，松手后恢复原始颜色触发一次
-              // 参数 bool，按下后true，恢复false
-              onHighlightChanged: (isClick) {
-                print(isClick);
-              }),
-          RaisedButton(
+            onPressed: () {  },
+              ),
+          ElevatedButton(
             onPressed: () {
               print("点击了凸起按钮");
             },
             child: Text("RaisedButton"),
           ),
-          RaisedButton.icon(
+          ElevatedButton.icon(
               onPressed: () {},
               icon: Icon(Icons.print),
               label: Text("带有icon的RaisedButton")),
-          RaisedButton(
-              child: Text("自定义RaiseButton", semanticsLabel: 'FLAT BUTTON 2'),
-              // 按钮颜色
-              color: Colors.cyan,
-              // 按钮亮度
-              colorBrightness: Brightness.dark,
-              // 高亮时的背景色
-              //highlightColor: Colors.yellow,
-              // 失效时的背景色
-              disabledColor: Colors.grey,
-              // 该按钮上的文字颜色，但是前提是不设置字体自身的颜色时才会起作用
-              textColor: Colors.white,
-              // 按钮失效时的文字颜色，同样的不能使用文本自己的样式或者颜色时才会起作用
-              disabledTextColor: Colors.grey,
-              // 按钮主题,主要用于与ButtonTheme和ButtonThemeData一起使用来定义按钮的基色,RaisedButton，RaisedButton，OutlineButton，它们是基于环境ButtonTheme配置的
-              //ButtonTextTheme.accent，使用模版颜色的;ButtonTextTheme.normal，按钮文本是黑色或白色取决于。ThemeData.brightness;ButtonTextTheme.primary，按钮文本基于。ThemeData.primaryColor.
-              textTheme: ButtonTextTheme.normal,
-              // 按钮内部,墨汁飞溅的颜色,点击按钮时的渐变背景色，当你不设置高亮背景时才会看的更清楚
-              splashColor: Colors.deepPurple,
-              // 抗锯齿能力,抗锯齿等级依次递增,none（默认),hardEdge,antiAliasWithSaveLayer,antiAlias
-              clipBehavior: Clip.antiAlias,
-              padding: EdgeInsets.only(
-                  bottom: 5.0, top: 5.0, left: 30.0, right: 30.0),
-              shape: drawShape(buttonShapeType),
-              // RaisedButton 的点击事件
-              onPressed: () {
-                // Perform some action
-                setButtonShapeType();
-              },
-              // 改变高亮颜色回掉函数，一个按钮会触发两次，按下后改变时触发一次，松手后恢复原始颜色触发一次
-              // 参数 bool，按下后true，恢复false
-              onHighlightChanged: (isClick) {
-                print(isClick);
-              }),
+          ElevatedButton(
+              onPressed: () {  },
+              child: Text("自定义RaiseButton:同TextButton", semanticsLabel: 'FLAT BUTTON 2'),
+              ),
           IconButton(
             icon: Icon(Icons.print),
             onPressed: () {
