@@ -6,7 +6,7 @@
 // ignore_for_file: lines_longer_than_80_chars
 // ignore_for_file: depend_on_referenced_packages
 
-import 'package:device_info_plus_web/device_info_plus_web.dart';
+import 'package:device_info_plus/src/device_info_plus_web.dart';
 import 'package:share_plus_web/share_plus_web.dart';
 import 'package:url_launcher_web/url_launcher_web.dart';
 
@@ -14,7 +14,7 @@ import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 // ignore: public_member_api_docs
 void registerPlugins(Registrar registrar) {
-  DeviceInfoPlusPlugin.registerWith(registrar);
+  DeviceInfoPlusWebPlugin.registerWith(registrar);
   SharePlusPlugin.registerWith(registrar);
   UrlLauncherPlugin.registerWith(registrar);
   registrar.registerMessageHandler();
