@@ -5,8 +5,8 @@
 import FlutterMacOS
 import Foundation
 
-import device_info_plus_macos
-import path_provider_macos
+import device_info_plus
+import path_provider_foundation
 import share_plus_macos
 import sqflite
 import url_launcher_macos
